User Authentication and search filter

+++ app/controllers/properties_controller.rb
@@ -1,17 +1,17 @@
 class PropertiesController < ApplicationController
-  before_action :authenticate_user!
+  # before_action :authenticate_user!
   before_action :load_config, only: :search_property
 
   def index
-    @properties = Property.all.order(rent: :asc)
+    @properties = Property.all.order(id: :asc)
 
     update_favorite_properties
 
-    render json: { data: @properties, count: @properties.count }, status: 200
+    render json: { data: @properties }, status: 200
   end
 
   def show
-    @property = Property.find(params[:id])
+    @property = Property.find_by(id: params[:id])
 
     render json: @property
   end
@@ -31,9 +31,11 @@ class PropertiesController < ApplicationController
   def update
     return unauthorized_message unless admin?
 
-    @property = Property.find(params[:id])
+    @property = Property.find_by(id: params[:id])
 
     if @property.update(property_params)
+      @property.is_favorite = false
+      @property.save!
       render json: @property, status: :ok
     else
       render json: @property.errors, status: :unprocessable_entity
@@ -43,7 +45,9 @@ class PropertiesController < ApplicationController
   def destroy
     return unauthorized_message unless admin?
 
-    property = Property.find(params[:id])
+    property = Property.find_by(id: params[:id])
+
+    return render json: {message: 'No Properties found'}, status: :not_found if property.nil?
 
     if property
       property.destroy
@@ -59,17 +63,42 @@ class PropertiesController < ApplicationController
     min_net_size = params[:min_net_size] || @config_data['min_net_size']
     max_net_size = params[:max_net_size] || @config_data['max_net_size']
     cities = params[:city]
+    property_type = params[:property_type]
     districts = params[:district]
     max_bedrooms = params[:max_bedrooms] || @config_data['max_bedrooms']
 
-    @properties = Property.residential_properties
-                          .within_rent_range(min_rent, max_rent)
-                          .within_net_size_range(min_net_size, max_net_size)
-                          .in_cities(cities)
-                          .in_districts(districts)
-                          .with_max_bedrooms(max_bedrooms.to_i)
+    @properties = Property.all # Initialize with all properties
 
-    properties = @properties.present? ? @properties : Property.all
+    # Filter by property type if property_type is present
+    if property_type.present?
+      @properties = @properties.residential_properties(property_type)
+    end
+    
+    # Filter by rent range if min_rent and max_rent are present
+    if (min_rent.present? && max_rent.present?) && (max_rent.to_i != 0)
+      @properties = @properties.within_rent_range(min_rent, max_rent)
+    end
+    
+    # Filter by net size range if min_net_size and max_net_size are present
+    if (min_net_size.present? && max_net_size.present?) && (max_net_size.to_i != 0)
+      @properties = @properties.within_net_size_range(min_net_size, max_net_size)
+    end
+    
+    # Filter by cities and districts if they are present
+    if cities.present?
+      @properties = @properties.in_cities(cities)
+    end
+    
+    if districts.present?
+      @properties = @properties.in_districts(districts)
+    end
+    
+    # Filter by max bedrooms if max_bedrooms is present
+    if max_bedrooms.present?
+      @properties = @properties.with_max_bedrooms(max_bedrooms.to_i)
+    end
+
+    properties = @properties.present? ? @properties : Property.all.where(property_type: 'residential')
 
     render json: properties
   end
@@ -85,16 +114,21 @@ class PropertiesController < ApplicationController
     end
 
     def property_params
-      params.require(:property).permit(:property_type, :country, :area, :net_size, :rent)
+      params.require(:property).permit(:title, :address, :net_size, :rent, :property_type, :country, :area, :number_of_living_rooms, :number_of_bathrooms)
     end
 
     def update_favorite_properties
-      return unless current_user&.favorite&.favorite_properties&.any?
-
-      current_user&.favorite&.favorite_properties&.each do |favorite_property|
-        property = Property.find(favorite_property.property_id)
-        property.is_favorite = true
-        property.save
+      if current_user&.favorite&.favorite_properties&.any?
+        current_user&.favorite&.favorite_properties&.each do |favorite_property|
+          property = Property.find(favorite_property.property_id)
+          property.is_favorite = true
+          property.save
+        end
+      else
+        Property.all.each do |property|
+          property.is_favorite = false
+          property.save
+        end
       end
     end
